rename dds test and join metadata-server into test-metadata
